pyproject.toml
[project]
name = "diarization-dataset"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "datasets[audio]==3.6.0",
    "ffmpeg>=1.4",
    "soundfile>=0.13.1",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "pyarrow>=14.0.0",
    "huggingface_hub>=0.20.0",
    "pydantic>=2.0.0",
]
